backend/chatbot-service/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
httpx==0.25.1
redis==5.0.1
motor==3.3.2

# NLP and ML dependencies
torch>=2.0.0
transformers==4.36.0
sentence-transformers==2.2.2
numpy==1.24.3
pandas==2.1.3
scikit-learn==1.3.2
nltk==3.8.1
spacy==3.7.2

# For intent classification and NER
datasets==2.15.0
accelerate==0.25.0
tokenizers==0.15.0

# Response generation
openai==1.3.8  # Optional for GPT integration
langchain==0.0.348
chromadb==0.4.18  # For vector storage

# API and async support
aiofiles==23.2.1
websockets==12.0
pydantic==2.5.2
pydantic-settings==2.1.0

# Monitoring and logging
prometheus-client==0.19.0
python-json-logger==2.0.7
pythonjsonlogger==2.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1